Delete pipeline code and move to workflow, change tests

+++ pymicmac/workflow/run_param_estimation.py
@@ -3,7 +3,7 @@ import os, argparse
 from lxml import etree
 from pymicmac import utils_execution
 
-def run(name, extension, configFile, mountPoint, onlyPrint):
+def run(name, extension, configFile, mountPoint, onlyPrint, noInit):
     rootPath = os.getcwd()
 
     e = etree.parse(configFile).getroot()
@@ -16,8 +16,9 @@ def run(name, extension, configFile, mountPoint, onlyPrint):
         if typeToLinkComponent != None:
             toLink.extend(typeToLinkComponent.text.strip().split())
 
-    # Initialize workflow folder (i.e. create link of images, homol folder and other files required by processing)
-    utils_execution.initPipelineFolder(name, extension, toLink)
+    if not noInit:
+        # Initialize workflow folder (i.e. create link of images, homol folder and other files required by processing)
+        utils_execution.initPipelineFolder(name, extension, toLink)
 
     # Change directory to workflow folder
     os.chdir(name)
@@ -32,15 +33,20 @@ def run(name, extension, configFile, mountPoint, onlyPrint):
         if componentName == "OriRedTieP":
             # From now one we use as Homol folder the one we just created
             if not onlyPrint:
+                if not os.path.isdir('HomolTiePRed'):
+                    raise Exception('OriRedTieP did not generate a folder with reduced set of points!')
                 os.system('rm Homol')
                 os.system('mv HomolTiePRed Homol')
         if componentName == "RedTieP":
             if "ExpSubCom=1" in componentOptions:
+                noodlesNumProc = mmComponent.find("noodlesParallel").text.strip()
                 noodlesExePath = os.path.abspath(os.path.join(os.path.realpath(__file__),'../../noodles/noodles_exe_parallel.py'))
                 command = 'python ' + noodlesExePath + ' -j ' + str(noodlesNumProc) + ' subcommands.json RedTieP_logs'
                 utils_execution.executeCommandMonitor('Noodles', command, mountPoint, onlyPrint)
             # From now one we use as Homol folder the one we just created
             if not onlyPrint:
+                if not os.path.isdir('Homol-Red'):
+                    raise Exception('OriRedTieP did not generate a folder with reduced set of points!')
                 os.system('rm Homol')
                 os.system('mv Homol-Red Homol')
 
@@ -56,11 +62,12 @@ def argument_parser():
     parser.add_argument('-c', '--configFile',default='', help='', type=str, required=True)
     parser.add_argument('-m', '--mountPoint',default='', help='', type=str, required=True)
     parser.add_argument('-p', '--onlyPrint', default=False, action='store_true')
+    parser.add_argument('--noInit', default=False, action='store_true')
     return parser
 
 if __name__ == "__main__":
     try:
         a = utils_execution.apply_argument_parser(argument_parser())
-        run(a.name, a.extension, a.configFile, a.mountPoint, a.onlyPrint)
+        run(a.name, a.extension, a.configFile, a.mountPoint, a.onlyPrint, a.noInit)
     except Exception as e:
         print(e)
